Make binary() output float32 on all backends (#3511)
_cpu_binary allocated np.empty(..., dtype=data.dtype), so binary()
returned the input dtype on numpy/dask+numpy while cupy/dask+cupy always
returned float32. Allocate float32 in _cpu_binary so all four backends
agree and so the NaN sentinel for non-finite cells is representable on
integer input. Matches _cpu_bin and the other classifiers.

Add cross-backend dtype regression tests.

+++ xrspatial/classify.py
@@ -48,7 +48,13 @@ def _available_memory_bytes():
 
 @ngjit
 def _cpu_binary(data, values):
-    out = np.empty(data.shape, dtype=data.dtype)
+    # Output float32 to match the cupy/dask+cupy backends (which always
+    # allocate 'f4') and the other classifiers, which route through _cpu_bin
+    # and likewise emit float32. Preserving the input dtype here made binary()
+    # the lone op whose result dtype diverged across backends (float64 on
+    # numpy/dask vs float32 on cupy) and could not hold the NaN sentinel for
+    # integer input.
+    out = np.empty(data.shape, dtype=np.float32)
     out[:] = np.nan
     rows, cols = data.shape
     for y in range(0, rows):
